#448 Add check for circular data references when resolving record id's
When resolving recordIds we now check if there are circular references to avoid infinite loops and hanging threads. If a circular reference is found a BadDataException is thrown.
This commit also fixes a problem where extra information about a problem was saved in the static ProblemType enum, which means this information was never removed and reused when a second problem occurred